[REFACT]: EndedChallenge 가져오는 API 호출 MyPageRepository로 이동
