+++ Projects/Domain/Repository/Interfaces/ChallengeRepository.swift
@@ -12,7 +12,6 @@ import Entity
 
 public protocol ChallengeRepository {
   func fetchPopularChallenges() -> Single<[ChallengeDetail]>
-  func fetchEndedChallenges(page: Int, size: Int) -> Single<[ChallengeSummary]>
   func fetchChallengeDetail(id: Int) -> Single<ChallengeDetail>
   func fetchMyChallenges(page: Int, size: Int) -> Single<[ChallengeSummary]>
   func fetchChallengeDescription(challengeId: Int) -> Single<ChallengeDescription>
