+++ Projects/Domain/Repository/Interfaces/MyPageRepository.swift
@@ -14,4 +14,5 @@ public protocol MyPageRepository {
   func fetchMyPageSummary() -> Single<MyPageSummary>
   func fetchVerifiedChallengeDates() -> Single<[Date]>
   func fetchFeedHistory(page: Int, size: Int) async throws -> PaginationResultType<FeedHistory>
+  func fetchEndedChallenges(page: Int, size: Int) async throws -> PaginationResultType<ChallengeSummary>
 }
